+++ plugin/src/main/java/com/atlassian/plugin/connect/plugin/capabilities/BeanToModuleRegistrar.java
@@ -13,9 +13,9 @@ import com.atlassian.plugin.Plugin;
 import com.atlassian.plugin.connect.modules.annotation.ConnectModule;
 import com.atlassian.plugin.connect.modules.beans.*;
 import com.atlassian.plugin.connect.modules.beans.builder.ConnectAddonBeanBuilder;
-import com.atlassian.plugin.connect.modules.beans.nested.ScopeName;
 import com.atlassian.plugin.connect.modules.util.ProductFilter;
 import com.atlassian.plugin.connect.plugin.capabilities.provider.ConnectModuleProvider;
+import com.atlassian.plugin.connect.plugin.capabilities.validate.AddOnBeanValidatorService;
 import com.atlassian.plugin.connect.plugin.descriptor.InvalidDescriptorException;
 import com.atlassian.plugin.connect.plugin.exception.ModuleProviderNotFoundException;
 import com.atlassian.plugin.connect.plugin.integration.plugins.DescriptorToRegister;
@@ -55,23 +55,24 @@ public class BeanToModuleRegistrar
     private final ProductAccessor productAccessor;
     private final ContainerManagedPlugin theConnectPlugin;
     private final ApplicationProperties applicationProperties;
-    private final WebHookScopeService webHookScopeService;
+    private final AddOnBeanValidatorService addOnBeanValidatorService;
 
     @Autowired
-    public BeanToModuleRegistrar(DynamicDescriptorRegistration dynamicDescriptorRegistration, PluginRetrievalService pluginRetrievalService, ProductAccessor productAccessor,
-                                 ApplicationProperties applicationProperties, WebHookScopeService webHookScopeService)
+    public BeanToModuleRegistrar(DynamicDescriptorRegistration dynamicDescriptorRegistration,
+            PluginRetrievalService pluginRetrievalService, ProductAccessor productAccessor,
+            ApplicationProperties applicationProperties, AddOnBeanValidatorService addOnBeanValidatorService)
     {
         this.dynamicDescriptorRegistration = dynamicDescriptorRegistration;
         this.productAccessor = productAccessor;
         this.applicationProperties = applicationProperties;
+        this.addOnBeanValidatorService = addOnBeanValidatorService;
         this.theConnectPlugin = (ContainerManagedPlugin) pluginRetrievalService.getPlugin();
         this.registrations = new ConcurrentHashMap<String, DynamicDescriptorRegistration.Registration>();
-        this.webHookScopeService = webHookScopeService;
     }
 
-    public void registerDescriptorsForBeans(Plugin plugin, ConnectAddonBean addon)
+    public void registerDescriptorsForBeans(Plugin plugin, ConnectAddonBean addon) throws InvalidDescriptorException
     {
-        requireScopesForWebHooks(plugin, addon);
+        addOnBeanValidatorService.validate(plugin, addon);
 
         BundleContext addonBundleContext = ((OsgiPlugin) plugin).getBundle().getBundleContext();
         AutowireWithConnectPluginDecorator connectAutowiringPlugin = new AutowireWithConnectPluginDecorator((AutowireCapablePlugin) theConnectPlugin, plugin, Sets.<Class<?>>newHashSet(productAccessor.getConditions().values()));
@@ -89,21 +90,6 @@ public class BeanToModuleRegistrar
         if (!descriptorsToRegister.isEmpty())
         {
             registrations.putIfAbsent(plugin.getKey(), dynamicDescriptorRegistration.registerDescriptors(plugin, descriptorsToRegister));
-        }
-    }
-
-    // don't leak content via web hooks to add-ons without permission to receive that content
-    private void requireScopesForWebHooks(Plugin plugin, ConnectAddonBean addon)
-    {
-        for (WebHookModuleBean webHookModuleBean : addon.getModules().getWebhooks())
-        {
-            ScopeName requiredScope = webHookScopeService.getRequiredScope(webHookModuleBean.getEvent());
-
-            if (!addon.getScopes().contains(requiredScope))
-            {
-                throw new InvalidDescriptorException(String.format("Add-on '%s' requests web hook '%s' but not the '%s' scope required to receive it. Please request this scope in your descriptor.",
-                                                                   plugin.getKey(), webHookModuleBean.getEvent(), requiredScope));
-            }
         }
     }