+++ plugin/src/main/java/com/atlassian/plugin/connect/plugin/installer/ConnectAddOnInstaller.java
@@ -2,6 +2,7 @@ package com.atlassian.plugin.connect.plugin.installer;
 
 import com.atlassian.plugin.Plugin;
 
+import com.atlassian.upm.spi.PluginInstallException;
 import org.dom4j.Document;
 
 /**
@@ -9,7 +10,7 @@ import org.dom4j.Document;
  */
 public interface ConnectAddOnInstaller
 {
-    Plugin install(String username, Document document);
+    Plugin install(String username, Document document) throws PluginInstallException;
 
-    Plugin install(String username, String capabilities);
+    Plugin install(String username, String capabilities) throws PluginInstallException;
 }