+++ plugin/src/test/java/it/com/atlassian/plugin/connect/rule/DisableDevMode.java
@@ -1,0 +1,12 @@
+package it.com.atlassian.plugin.connect.rule;
+
+/**
+ * Disable dev mode for the duration of a test.
+ */
+public class DisableDevMode extends OverrideSystemProperty
+{
+    public DisableDevMode()
+    {
+        super("atlassian.dev.mode", Boolean.FALSE.toString());
+    }
+}